+++ CHANGELOG.md
@@ -5,6 +5,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- You can now configure this plugin to run as a platform
+
+### Changed
+- Platform is now also the preferred way to use the plugin, when still using as an accessory a warning will be shown.
+- Updated README
+
+Note: All users should switch to use the plugin as a platform since the accessory way will be removed soon. In order to switch to platform you just simply need to adjust your config.json as the example in the README.
+
 
 ## [1.2.1] - 2020-07-26
 ### Changed